Fix WaitForNavigation within the same document
Previously WaitForNavigation assumed that a LifecycleEvent would always
be fired, but this is not the case for navigation within the same
document (e.g. via anchor links or the History API), so in those cases
the call would timeout after 30s.

The fix simply checks if we received a new document, otherwise it skips
waiting for the LifecycleEvent. Even if that wait didn't time out, it
would've failed with a nil pointer panic accessing event.newDocument.request.

Closes #226