Fix waitForFunction()
This is essentially a re-implementation of `waitForFunction()` since the
previous version was broken in many ways. This required changes in
`injected_script.js` where I hopefully didn't break anything else.
I tried to split it into several commits but the changes were too messy
and it would be very difficult at this point.

In summary this:
- Fixes `waitForFunction()` option parsing.
- Ensures all three polling variants work: `requestAnimationFrame`, DOM
  mutation and interval.
- Ensures we properly time out in all cases and return an informative
  error message.
- Makes `waitForFunction()` an async method.

Fixes #291