Rename k6 pkg to k6ext
Due to a bug in xk6 (grafana/xk6#45), we can't
have a directory or a file with a name k6.

As discussed internally, @imiric suggested renaming the k6 pkg to k6ext:
"k6ext follows the k6 convention of doing the same: errext, netext,
etc., so let's stick to that.".