Fix concurrent usage of goja.Runtime in waitForNavigation
This is a functional change from before, since the iteration won't be
interrupted by a timeout error anymore, and the promise will be rejected
instead. This might be an acceptable tradeoff for using a lock, but it
would be better if there was a way of interrupting the iteration if the
promise is rejected.

Fixes #506