Remove go and use chan to keep CDP msg order
The order of the CDP messages can quite easily be lost after being
received and handed to the handlers that are listening in on the
events. To alleviate this ordering issue, remove the goroutine that is
used to place the event on the listening handlers, and instead use a
buffered channel. The buffered channel avoids a deadlock where by an
event handler could be busy emitting another event, while the emit
method is busy trying to pass an event to the same handler.