WIP Don't cancel the browser context in the initEvents() goroutine defer
This doesn't belong here, because:

a) It was prematurely killing the browser process, which resulted in
`signal: killed` debug messages.

b) b.evCancelFn was assigned, but never called. The defer actually calls
b.cancelFn, which cancels the parent context, and might've been a bug.

c) If the goroutine returns it's either because the context is
cancelled, or we received the EventConnectionClose event. If the
parent context is cancelled, there's no need to cancel the child one.
And we can trigger didLoseConnection() directly in the
EventConnectionClose case.