Get DevTools URL from data dir file instead of by parsing stdout
This is slightly more elegant than parsing stdout, but more importantly,
shouldn't cause race conditions where we miss the "DevTools listening on" line,
since it doesn't start a goroutine.

A drawback is that it could take slightly longer to return the URL
because it's reading from disk and does several read attempts with delays,
but in my tests the difference is negligible. In total, it allots 500ms
to read the file, which should hopefully be plenty even on IO-starved
machines (VMs, CI, etc.). IMO if this takes more time than that then
it's likely the environment will not be fit for running a test.

Fixes #491