Add lifecycle event tests
Some of tests lock in the load and DOMContentLoaded lifecycle event
behaviour when used with page.goto. The other tests are the similar
except they ensure that all three types of lifecycle events work as
expected when we have a main frame and sub frame -- i.e. we should
wait until both the sub and main frame receive the lifecycle events
from the browser.