Fix deadlock with executionContextMu
executionContextMu isn't needed to be locked and unlocked in this
(WaitForFunction) method as it doesn't r/w to executionContext or
documentHandle. This function later calls an unexported method with the
same name, which does require access to executionContext.

A deadlock was occurring later in waitForExecutionContext since
this method was holding on to the lock.

Closes: https://github.com/grafana/xk6-browser/issues/635