Add cdp perf timeline event to capture web vitals
We're testing the CDP performance timeline event. It is able to
retrieve the LCP and LS metrics via the CDP perf timeline event.

Unfortunately LCP only returns the time when the largest contentful
paint occurred but not when it started, so we can't determine how long
it took.

LS does return us a score per frame. We would need to calculate the
score for the whole page (add up the scores for all frames) and then
work out the cumulative score.