Add an env var to disable test runs
This flag can stop tests runs that shouldn't occur, along with
another messaging env var that override the default messaging.

We're not using os.Exit to exit the process as that will prevent k6
from shutting down cleanly. Instead we throw an error that k6 can
work with and cleanly shut itself down, also allowing it to write an
appropriate exit code the other systems can understand and work with.

Note that this this throw needs to be done in NewModuleInstance.

Closes: https://github.com/grafana/xk6-browser/issues/664