Remove await method from testBrowser
Now that promises have been moved to a level above the business
logic, we can remove the need to run things in promises and assume
(in most cases) that things will run synchronously. Where there's
still a need to run things in parallel we can use the run method on
testBrowser.