Add the process register
This process register will store all the currently running browser
pids that xk6-browser has started in the current test iteration. It's
main use is to be able to quickly and easily shutdown these processes
if a panic occurs. This negates the need to store the pid in the
context.

This also solves an issue where if more than one browser was started,
only the last one could be forced to shutdown.