Fix the register to work with the tests
Although the register works for a normal test run with multiple
browser where a panic will warrant the closure of all the browser
instance, this is not the same for unit tests.

Unit tests share the same register, which is a problem since if
one test run panics, then all currently open browser instances will
close.

To fix this, register is now aware of interval ids which is a way for
unit tests to isolate panics within their own test. Normal running of
xk6-browser doesn't require an interval id.