Fix data race bugs - not one hundred percent
There is a data race somewhere in the metrics sampling channel.

If I enable the logs, the race goes away :^ This is the nature of the
heisenbugs.

Adding Context.Done checking in several places also helps.

One other discovery is that skip canceling the browser context in
BrowserType.launch() makes things more stable. This might be something
to do with that.

After these fixes only 10% of the runs fail. I think we can live with
that. Or we will need to compile a k6browser like we do on CI E2E and
then run the tests by executing it in parallel in Go.