Remove --no-sandbox as default browser flag
This increases the security by default when launching chrome browser.
This was only required when running xk6-browser as root, which mainly
happened when being executed in a Docker container. Now if xk6-browser
is run as root, due to the removal of this parameter, an error message
like "Running as root without --no-sandbox is not supported" will be
shown. Still end users will be able to disable this (run without
sandbox) by overwritting the default chrome args.