Add method on bCtx so page can apply init scripts
When a new page is created, it needs to add the init scripts that the
browser context has had applied to it.

Instead of retrieving the list of scripts to apply in the page, we're
sending the current page to the browser context, and it will loop over
the scripts it has in evaluateOnNewDocumentSources.