Update nil session log warn to debug in page attach
Rationale:

I suggested this earlier on to detect nil sessions. But now, this
warning is outdated because when users running multiple instance/VU
tests, they will see dozens/hundreds lines of warnings.

The core reason we receive a lot of these warnings is:

We need to correctly handle the CDP messages while sending and receiving
them (the order of them).
https://github.com/grafana/xk6-browser/pull/848