Add request failed metric
This metric is predefined in k6. When a request completes we want to
measure how many of these failed. This is based on the status code of
the response. If the code is outside of the 200-399 range, it is deemed
to have failed. The default status code range was taken from k6.

The opposite is also possible to determine, we can work out which
requests succeeded since the value of the metric will be 0.