Refactor contexts to context in browser
To enable us to predicate and have a better idea of how much resource
a test run will require, we want to restrict the number of browser-
Contexts per vu to be one. This will ensure that a single test
scenario which requires 1 vu and 50 iterations could work with a single
browser and a new browserContext per iteration.

Users will need to close the existing browserContext before creating
a new one. We feel that this restriction along with scenarios should
still work for majority of our users.