Add restriction on one context per iteration
This change will restrict the creation and use of a single browser
context per iteration. Once the browser context has been closed, a new
one cannot be created, setup or any other actions against. This
essentially ends the browser test for the iteration, but allows the
user to run other k6 tests still.

The reason for this change is:
1. We want to prevent the user to work with a single browser context
per iteration, and therefore restrict the scenario to one type of
browser context. If the user wishes to work with different types of
browser contexts, then they should work with scenarios to create a new
test that changes the type of browser context which will affect the
test. This helps separate the concerns of the two types of browser
context.
2. Creating and closing browser contexts while also running async APIs
will cause errors and issues for the user. They are likely to be a
confusing experience. We want to avoid this.
3. The browser module processes background async operations when APIs
are called. If a browser context is closed, this will affect how we
process those background operations which will inadvertently affect the
test and the results.