Fix deadlock by removing lock in frameNavigated
In certain navigations, the current frame may have child frames. In
these cases, the child frames need to be removed. frameNavigated holds
on to a framesMu lock for the whole duration of the frameNavigated.

removeFramesRecursively, which is what removes the child frames, needs
the same lock so that it can amend the synchronised variable.

It's also worth noting that this also prevent frameNavigated holding on
to a lock when it calls ChildFrames, which requires another lock. It's
generally unsafe to hold onto more than one lock at a time.