Add a test to lock in behaviour for nav to iframe
This test actually tests that we can navigate to a page with iframes,
click on a link, which navigates to another page with iframes. Once
at the second page it will end the test and we can assert. Without the
fix in the prior commit, this would panic on a deadlock.

NOTE: This only works when we have headless mode on.