Report all web vital metric changes
With this change, the Web Vital callbacks will be called for all
changes. instead of waiting for them to complete.

See this comment:
If the `reportAllChanges` configuration option is set to `true`, the
`callback` function will be called as soon as the value is initially
determined as well as any time the value changes throughout the page
lifespan.

https://github.com/GoogleChrome/web-vitals/blob/5973d51be39ab0020268aa2397ead0962eedf4c2/src/onCLS.ts#L43-L46