Add K6_BROWSER_EXECUTABLE_PATH in ci
The github runners have both chromium and google chrome installed.
Google chrome is kept up to date whereas chromium isn't. We want to
work with the latest version of google chrome (for now, in the future
we may want to install the exact version we want to work with).

We've had issues working with chromium in the runners where the
application crashes on startup.