Add a new EvaluateWithContext
This new EvaluateWithContext will allow us to pass in a context with a
timeout. For issue #971, we're seeing that often the evaluate to hide
the page on page.close ends up in a deadlock (waiting for the response
from the chrome browser after executing the CDP evaluate command). We
are not sure of the root cause of the issue, but this fix will help
prevent this deadlock allow the test run to complete in a timely
manner without it causing a time out error.