Use EvaluateWithContext from page.close
Now we can use the EvaluateWithContext from page.close when we call
evaluate to hide the page. We're going to rely on the default timeout
for now, and refine the timeout if we see this occurring too often --
if we see this occur too often this will affect the test result
depending on the default timeout (which is 30 seconds unless
overridden by the user).