Add a msgId type that will help create unique ids
Currently the msgIds clash from one session to another as well as with
messages that are sent without a session on the connection event loop.

Generally this isn't an issue, but recently we have found that chrome
sometimes returns an error without the sessionId that was in the
original request msg. When this occurs, we route those error responses
to handlers that are waiting on the connection event loop. If there's
a clash in msgId then the wrong handler will work with the wrong
response, which could result in incorrect behaviour.

This change will enforce a unique msgId at the connection level, which
should avoid such scenarios.