Refactor test to write to separate dir
There will be two temp clean dir tests. We need them to not look at
what each other are creating and cause the test to pass/fail. So the
current test will need to write to a new dir that is different to what
the new test will write to.