Add api.Cookie type
We can use this type in our code and tests and pass to clients. The
network.Cookie contains an internal fields and we don't want to use it.

This type provides type safety and representation. Instead of a map type
which its fields are of the same type, not type-safe, and not useful or
effective for representing a concept like a cookie.

Declaring this in the api package. So we can have less verbosity in the
rest of the codebase. The downside is the api package dependency.
If we declared this in the common package, there would be circular
dependency issues.