Refactor to start/end trace on k6 events
This change aligns with the new event system from k6 which informs
the browser module when the iteration starts, end and when the test run
ends. This allows the browser module to start a trace on iterStart, end
a trace on iterEnd, and shutdown the traceProvider when the test run
completes.

Note: on iterStart the background context is used instead of the ctx
that is a member variable of tracesRegister. The reason for this is
so that we're not bound by the vu context.