Add a k6 object script file
To be consistent with how we work with other injected scripts, the k6
object is being moved to its own js file. This also helps us avoid
working with globals.

It's worth noting that the k6 object will need to be mutable at some
stage to add test specific unique metadata in the future.

Resolves: https://github.com/grafana/xk6-browser/pull/1007#discussion_r1311155144