Add cookie with expiration test
Also fixes a bug in the test code:
tt.wantCookiesToSet, len(tt.cookies),

This was comparing the cookies to set to wantCookiesToSet. But, what we
want to compare with is the cookies that we get from the browser
context. While testing with expired cookies, we can give three cookies,
one of them should be expired and not to be found in the browser
context.