Add a liveSpan type for tracking page navigations
We need to be able to track async events i.e. web vitals. We
eventually want to be able to track the web vital against the API call
that initiated the web vital measurement. This still requires quite a
bit of investigative work.

For now, we're going to track the web vitals across the whole page
navigations. If the test navigates to test.k6.io, then we need to track
the start of the navigation until the next navigation OR the end of the
test OR the end of the whole test run.

This type will help us track the current live page navigation spans.