Add a new function to track the PageNavigation
A page navigation will be a child of the root (iteration) trace. It
will start when a navigation starts and end when the next navigation
OR the test iteration ends OR when the whole test run ends.

Since there isn't an easy way to propagate the new context created when
the span is created, we will for now store the context in a local map.

When an API span is needed it will need to pass in the targetID it is
associated with to retrieve the context and add itself as a child span
of the PageNavigation span.