Add trace in onFrameNavigated
When a frame navigates we want to be able to start and stop a
PageNavigation span. This span will be used to associate async event
(i.e. web vitals) to the whole page but not currently to the API calls
that may have attributed to the measurements to be taken.