Add a temporary method to evaluate js code
When working with Evaluate when trying to inject the spanID into the
page, it kept erroring with "execution context changed; most likely
because of a navigation".

After debugging this, it was found that not passing in the exec context
id prevented such an error. For now I've created a temporary method
to evaluate js on a page without the exec context id.