Add a new method to track API call as spans
We want to be able to associate the API calls as spans. Since there
is currently no easy way to propagate a PageNavigation context back
to the mapping layer, we're having to track them with the targetID and
keep a record of them in the otel package.

This new method requires the targetID to be able to create a span on
the correct PageNavigation span, so that an API call is easily attached
as a child to the PageNavigation span.