Refactor browser spans to work with TraceAPICall
Refactoring the current calls to create a span to work with
TraceAPICall. We're also working with b.ctx instead of the context
that has been passed in since b.ctx has the iteration trace id
embedded into it.

Since there won't be a PageNavigation span at this point when the api
is called, instead the span will be linked to the iteration trace.