Refactor browserContext trace creation
This commit refactors the code to work with the TraceAPICall so that it
is correctly associated to the correct span.

We're working with b.ctx here instead of the passed in context since
b.ctx has the trace span context embedded in it.

In this case, since there won't be a new PageNavigation span, it will
be associated to the iteration span.