Fix deadlock which occur in integration test
In one particular integration test it deadlocks when the evaluate is
called in onFrameNavigated (TestFrameDismissDialogBox). What i think is
happening is that the onEventJavascriptDialogOpening event is waiting
to be handled but the current goroutine is busy performing the
evaluate CDP request and then waiting for a response, which is probably
being queued up behind the onEventJavascriptDialogOpening event.

It would be good to fire and forget the evaluate call or be able to put
it on the event loop to free up the frame session goroutine.