Update mapping to promisify waitForEvent
waitForEvent is useless without it being promisified. This changes
that so that waitForEvent returns a promise, allowing the user to set
a waitForEvent and then call further APIs which will unblock the
waitForEvent when possible (only for new page creations and if the
predicate function is set and returns a truthy value).