Remove browser close event
This event isn't being fired when the browserContext is closed. If its
left in it just adds confusion to anyone reading the code.

Since close events all use the keyword "close" (e.g. when a page closes
it emits a "close" event), we don't want the wrong type of a close
event causing the wait to unblock.