Refactor the predicate call for waitForEvent
This change refactors the predicate call so that it happens on the
taskqueue which prevents unexpected behaviour such as nil issues and
race conditions within goja.

Since the task on taskqueue will be called sometime in the future, we
need to wait for the response from the predicate and channel that
response correctly back to the appropriate existing channels out and
errOut.