Update taskqueue creation
It's a smell to have the page's taskqueue created from the
browserContext when the waitForEvent API is used. Instead it's more
maintainable to have the queue creation occur when the page is created.

The downside is that now the user MUST close the page otherwise it will
cause the test to hang indefinitely.