Initialize traces registry on VU iterStart
The traces registry has to be held per VU scope, as it stores traces
indexed per iteration, which are only unique per VU. Therefore keep it
under browser registry so we can use it to handle traces generation
based on iter events.
Because VU.State is nil when the browser registry is initialized, we
have to initialize the traces registry on the first VU iteration so we
can get access to the k6 TracerProvider. This is done within a sync.Once
so we ensure the traces registry is only initialized once per VU.