Inject spanID on frame navigated
In order to be able to correlate the Web Vitals events to the navigation
span to which they belong to, we have to add a reference to that span in
the page context, so it can be retrieved by our Web Vitals script that
is executed in the page before pushing the WV metrics to k6 browser
through the set up binding.

The JS code to set the spanID in the page context is executed through
the EvaluateGlobal method as otherwise it would be executed in the
current frame context. The problem is that, because we are processing
a navigation event, the frame context has actually already changed in
the browser but our event parsing goroutine will not process that until
we exit the onFrameNavigated method and receive the
onExecutionContextCreated event. This would result in an error when
evaluating the JS code due to "execution context changed".
In this case, because the JS code sets a property in the global object
(window), it is OK to just execute it in the global context.

Additionally the EvaluateGlobal method has to be called in a separate
goroutine to avoid deadlocking the event parsing goroutine when the
frame loading generates CDP events that have to be acted on for the
loading process to continue, such as when the frame contains a JS
initiated dialog (alert, confirm, prompt, or onbeforeunload) which we
have to parse and respond with an explicit accept/dismiss action (see
onEventJavascriptDialogOpening).